backend/pyproject.toml
[project]
name = "multi_agent_stock_analyst"
version = "0.1.0"
description = "Multi-agent stock analyst using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai-tools>=1.14.1",
    "crewai[tools]==1.14.1",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.4",
    "newsapi-python>=0.2.7",
    "pydantic>=2.11.10",
    "uvicorn>=0.44.0",
    "websockets>=16.0",
    "yfinance>=1.2.2",
]

[project.scripts]
multi_agent_stock_analyst = "multi_agent_stock_analyst.main:run"
run_crew = "multi_agent_stock_analyst.main:run"
train = "multi_agent_stock_analyst.main:train"
replay = "multi_agent_stock_analyst.main:replay"
test = "multi_agent_stock_analyst.main:test"
run_with_trigger = "multi_agent_stock_analyst.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
